CTCM3/requirements.txt
torch>=2.2
torchvision>=0.18
einops>=0.7
tqdm>=4.66
pyyaml>=6.0
clean-fid>=0.1
